backend/requirements.txt
asgiref==3.8.1
Django ==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg ==1.21.8
inflection==0.5.1
packaging==24.2
pip==22.3.1
pytz==2024.2
PyYAML==6.0.2
setuptools==65.5.0
sqlparse==0.5.3
uritemplate==4.1.1